# frontend/schemas/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SlotMath.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5152/"
    }
  ],
  "paths": {
...
    "/api/runs/{id}/measurements/reference/comparison": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExactLawComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetainedReferenceOfExactLawComparisonReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/measurements/reference/distribution": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiniteModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetainedReferenceOfFiniteModelReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/measurements/reference/markov": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkovModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetainedReferenceOfMarkovModelReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/measurements/reference/planning": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplePlanReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/measurements/reference/comparison": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExactLawComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExactLawComparisonReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/measurements/reference/distribution": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiniteModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiniteModelReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/measurements/reference/markov": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkovModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkovModelReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/evidence": {
      "get": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEvidence"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/measurements/reference": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnumerationBudget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphMeasurementReference"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/measurements/schema": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementSchemaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnumerationBudget": {
        "type": "object",
        "properties": {
          "maximumPaths": {
            "type": "integer",
            "format": "int32",
            "default": 10000
          },
          "maximumOperations": {
            "type": "integer",
            "format": "int32",
            "default": 250000
          },
          "maximumFrontier": {
            "type": "integer",
            "format": "int32",
            "default": 2048
          }
        }
      },
...
      "ExactLawComparisonReport": {
        "required": [
          "equal",
          "leftMean",
          "rightMean",
          "meanDifference",
          "totalVariation",
          "cdfDistance",
          "support",
          "unit",
          "assumptions"
        ],
        "type": "object",
        "properties": {
          "equal": {
            "type": "boolean"
          },
          "leftMean": {
            "type": "string"
          },
          "rightMean": {
            "type": "string"
          },
          "meanDifference": {
            "type": "string"
          },
          "totalVariation": {
            "type": "string"
          },
          "cdfDistance": {
            "type": "string"
          },
          "support": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RationalLawDifference"
            }
          },
          "unit": {
            "type": "string"
          },
          "assumptions": {
            "type": "string"
          },
          "authoredInputSha256": {
            "type": [
              "null",
              "string"
            ]
          },
          "algorithmVersion": {
            "type": [
              "null",
              "string"
            ]
          },
          "coreBinarySha256": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ExactLawComparisonRequest": {
        "required": [
          "left",
          "right"
        ],
        "type": "object",
        "properties": {
          "left": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RationalOutcome"
            }
          },
          "right": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RationalOutcome"
            }
          },
          "unit": {
            "type": "string",
            "default": "value units"
          }
        }
      },
...
      "FiniteModelRequest": {
        "required": [
          "outcomes"
        ],
        "type": "object",
        "properties": {
          "outcomes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RationalOutcome"
            }
          },
          "cost": {
            "type": "string",
            "default": "1"
          },
          "prunedMass": {
            "type": "string",
            "default": "0"
          },
          "provenMaximum": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "MarkovModelRequest": {
        "required": [
          "transientMatrix",
          "rewards"
        ],
        "type": "object",
        "properties": {
          "transientMatrix": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "rewards": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "initialState": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "stationary": {
            "type": "boolean",
            "default": false
          },
          "costs": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "MeasurementSchema": {
        "required": [
          "points",
          "fields"
        ],
        "type": "object",
        "properties": {
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeasurementPoint"
            }
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeasurementField"
            }
          }
        }
      },
...
      "RetainedReferenceOfExactLawComparisonReport": {
        "required": [
          "runId",
          "report",
          "retention"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "report": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExactLawComparisonReport"
              }
            ]
          },
          "retention": {
            "$ref": "#/components/schemas/DiagnosticRetention"
          }
        }
      },
      "RetainedReferenceOfFiniteModelReport": {
        "required": [
          "runId",
          "report",
          "retention"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "report": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FiniteModelReport"
              }
            ]
          },
          "retention": {
            "$ref": "#/components/schemas/DiagnosticRetention"
          }
        }
      },
      "RetainedReferenceOfMarkovModelReport": {
        "required": [
          "runId",
          "report",
          "retention"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "report": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MarkovModelReport"
              }
            ]
          },
          "retention": {
            "$ref": "#/components/schemas/DiagnosticRetention"
          }
        }
      },
...
      "SamplePlanReport": {
        "required": [
          "allocatedAlpha",
          "normalCritical",
          "approximateMeanRounds",
          "timeUniformBoundedMeanRounds",
          "eventDetectionRounds",
          "zeroObservedEventUpper",
          "meanBudgetSufficient",
          "detectionBudgetSufficient",
          "assumptions"
        ],
        "type": "object",
        "properties": {
          "allocatedAlpha": {
            "type": "number",
            "format": "double"
          },
          "normalCritical": {
            "type": "number",
            "format": "double"
          },
          "approximateMeanRounds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "timeUniformBoundedMeanRounds": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "eventDetectionRounds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "zeroObservedEventUpper": {
            "type": "number",
            "format": "double"
          },
          "meanBudgetSufficient": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "detectionBudgetSufficient": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "assumptions": {
            "type": "string"
          },
          "authoredInputSha256": {
            "type": [
              "null",
              "string"
            ]
          },
          "coreBinarySha256": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SamplePlanRequest": {
        "type": "object",
        "properties": {
          "confidence": {
            "type": "number",
            "format": "double",
            "default": 0.95
          },
          "errorFamilySize": {
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          "precision": {
            "type": "number",
            "format": "double",
            "default": 0.001
          },
          "variance": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "lowerBound": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "upperBound": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "eventProbability": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "detectionPower": {
            "type": "number",
            "format": "double",
            "default": 0.95
          },
          "roundBudget": {
            "type": "integer",
            "format": "int64",
            "default": 100000
          }
        }
      },
...
    }
  },
...
}